requirements.txt
Flask==3.0.0
opencv-python==4.8.0.74
Pillow==10.0.0
python-dotenv==1.0.0
Werkzeug==3.0.1
pytube==15.0.0
moviepy==1.0.3
reportlab==4.0.4
fpdf==1.7.2
Flask-WTF==1.2.1
requests==2.31.0
beautifulsoup4==4.12.2
